feat: Add support for generating resource IDs in NamingConvention class
This commit adds support for generating resource IDs in the `NamingConvention` class. It introduces a new method `GenerateResourceId` that takes a resource type as input and returns the generated resource ID. The method uses a dictionary of resource type generators to generate the resource ID based on the provided inputs.

The commit also includes some minor changes, such as removing unused variables and fixing formatting issues.