feature: add-timestamps-handling (#6)
* feature: add-timestamps-handling

This commit adds timestamp handling to the code. It includes changes in the `AzureStack.cs`, `CoreComponents.cs`, `DataPlatform.cs`, `AzKeyVault.cs`, and `AzResourceGroup.cs` files. The changes involve adding timestamps to various tags and resource properties, such as creation date and created by information.

* feat: Update AzureStack class and Utils class

Fixes #7
- Updated the AzureStack class by adding a namespace declaration and fixing formatting issues.
- Updated the Utils class by removing duplicate nested class declaration and fixing formatting issues.


* feat: Add data and stuff to README

This commit adds a new section titled "Data and stuff" to the README file. It includes an image and some introductory text about the purpose of the project.

fix: Fix formatting in README

This commit fixes some formatting issues in the README file, including alignment and spacing.

refactor: Update Utils class to be static

This commit updates the `Utils` class to be static instead of internal. This change allows for easier access to its methods without needing an instance of the class.

fix: Add conditional logic for role assignments in AzKeyVault.cs

This commit adds conditional logic for role assignments in the `AzKeyVault` class. Depending on whether a user or service principal is specified, it assigns different roles accordingly.


Remove some code smell

- Remove unused imports and variables in CoreComponents.cs
- Refactor variable names in CoreComponents.cs for clarity
- Update date formatting in AzResourceGroup.cs

* feat: Add support for generating resource IDs in NamingConvention class

This commit adds support for generating resource IDs in the `NamingConvention` class. It introduces a new method `GenerateResourceId` that takes a resource type as input and returns the generated resource ID. The method uses a dictionary of resource type generators to generate the resource ID based on the provided inputs.

The commit also includes some minor changes, such as removing unused variables and fixing formatting issues.

* Refactor DataPlatform class and AzureStack class

- Remove StackConfig parameter from DataPlatform constructor
- Remove unused variables in DataPlatform class
- Update references to NamingConvention and tags in DataPlatform class
- Update references to NamingConvention and CloudProvider in AzureStack class

* feat: Add AzKeyVault class with methods for managing secrets and role assignments

This commit adds a new `AzKeyVault` class to the codebase, which provides methods for managing secrets and role assignments in Azure Key Vault. The `WithSecret` method allows adding a secret to the key vault, while the `WithSecretsReader` method enables granting read access to specified users or service principals. Additionally, the `WithKeyVaultSecretsAdmins` method allows assigning contributors as administrators of key vault secrets. These changes aim to enhance the functionality and flexibility of working with Azure Key Vault in the application.

* Update README.md and StringExpressions.cs

- Commented out the status badge in README.md
- Changed class modifier from "public" to "public static" in StringExpressions.cs

---------

Co-authored-by: Merca Ovnerud <merca@users.noreply.github.com>